# contracts/openapi/identity.yaml
...
openapi: 3.1.0
info:
  title: TogetherGo identity
  description: Accounts, authentication and RS256 token issuance. The private signing key exists only
    in this service; every other service verifies tokens against /.well-known/jwks.json.
  version: 0.1.0
paths:
...
  /api/auth/register:
    post:
      tags:
      - auth
      summary: Create an account and sign in
      operationId: register_api_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/login:
    post:
      tags:
      - auth
      summary: Exchange credentials for a token pair
      operationId: login_api_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/refresh:
    post:
      tags:
      - auth
      summary: Rotate a refresh token
      operationId: refresh_api_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/logout:
    post:
      tags:
      - auth
      summary: Revoke a refresh token chain
      operationId: logout_api_auth_logout_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AuthResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        expires_in:
          type: integer
          title: Expires In
        user:
          $ref: '#/components/schemas/PrivateProfile'
      type: object
      required:
      - access_token
      - refresh_token
      - expires_in
      - user
      title: AuthResponse
      description: "POST /api/auth/register and /login \u2014 pair plus the account."
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          maxLength: 256
          minLength: 1
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
    LogoutRequest:
      properties:
        refresh_token:
          type: string
          maxLength: 512
          minLength: 1
          title: Refresh Token
      type: object
      required:
      - refresh_token
      title: LogoutRequest
...
    PrivateProfile:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          format: email
          title: Email
        full_name:
          type: string
          title: Full Name
        bio:
          anyOf:
          - type: string
          - type: 'null'
          title: Bio
        birth_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Birth Date
        age:
          anyOf:
          - type: integer
          - type: 'null'
          title: Age
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        photo_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Photo Url
        rating_avg:
          anyOf:
          - type: number
          - type: 'null'
          title: Rating Avg
        rating_count:
          type: integer
          title: Rating Count
        is_active:
          type: boolean
          title: Is Active
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - email
      - full_name
      - rating_count
      - is_active
      - created_at
      - updated_at
      title: PrivateProfile
      description: 'GET /api/users/me, and the `user` of an auth response.


        The only shape that carries email, phone and birth date. Never returned

        for anyone but the authenticated caller themselves.'
...
    RefreshRequest:
      properties:
        refresh_token:
          type: string
          maxLength: 512
          minLength: 1
          title: Refresh Token
      type: object
      required:
      - refresh_token
      title: RefreshRequest
    RefreshResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        expires_in:
          type: integer
          title: Expires In
      type: object
      required:
      - access_token
      - refresh_token
      - expires_in
      title: RefreshResponse
      description: "POST /api/auth/refresh \u2014 new pair, no profile."
    RegisterRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          maxLength: 256
          minLength: 10
          title: Password
        full_name:
          type: string
          maxLength: 64
          minLength: 1
          title: Full Name
        birth_date:
          type: string
          format: date
          title: Birth Date
        bio:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Bio
      type: object
      required:
      - email
      - password
      - full_name
      - birth_date
      title: RegisterRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      description: RS256 access token
      scheme: bearer
    APIKeyHeader:
      type: apiKey
      description: Shared secret; must equal INTERNAL_API_TOKEN.
      in: header
      name: X-Internal-Token
servers:
- url: http://localhost:8080
  description: gateway
- url: http://identity:8001
  description: direct, internal network only